refactor: 파일 분리, 모듈화
