+++ src/App.ts
@@ -1,0 +1,9 @@
+import CounterAndMeow from './components/CounterAndMeow';
+
+const App = () => `
+  <div>
+    ${CounterAndMeow()}
+  </div>
+  `;
+
+export default App;
