+++ src/core/MyReact.ts
@@ -1,0 +1,48 @@
+import debounceFrame from '../utils/debounceFrame';
+
+interface Options {
+  currentStateKey: number;
+  renderCount: number;
+  states: any[];
+  root: Element | null;
+  rootComponent: any;
+}
+
+export default function MyReact() {
+  const options: Options = {
+    currentStateKey: 0,
+    renderCount: 0,
+    states: [],
+    root: null,
+    rootComponent: null,
+  };
+
+  function useState(initState: any) {
+    const { currentStateKey: key, states } = options;
+    if (states.length === key) states.push(initState);
+
+    const state = states[key];
+    const setState = (newState: any) => {
+      states[key] = newState;
+      _render();
+    };
+    options.currentStateKey += 1;
+    return [state, setState];
+  }
+
+  const _render = debounceFrame(() => {
+    const { root, rootComponent } = options;
+    if (!root || !rootComponent) return;
+    root.innerHTML = rootComponent();
+    options.currentStateKey = 0;
+    options.renderCount += 1;
+  });
+
+  function render(rootComponent: any, root: Element | null) {
+    options.root = root;
+    options.rootComponent = rootComponent;
+    _render();
+  }
+
+  return { useState, render };
+}
