+++ src/utils/debounceFrame.ts
@@ -1,0 +1,7 @@
+export default function debounceFrame(callback: FrameRequestCallback) {
+  let nextFrameCallback = -1;
+  return () => {
+    cancelAnimationFrame(nextFrameCallback);
+    nextFrameCallback = requestAnimationFrame(callback);
+  };
+}
